requirements.txt
openai
python-dotenv
pypdf
scikit-learn
